Creating Monorepo with TurboRepo
Creating a SaaS project with Next.js, covering permissioning fundamentals such as RBAC and ABAC. Using Node.js, Fastify and Prisma on the back-end, and Next.js 14 on the front-end with Server Components and Server Actions. Implementing the monorepo to keep back-end and front-end together. TurboRepo will be used to speed up processes, avoiding unnecessary executions. Making adjustments to packages and configurations to start the NextSAS-RBAC project.