Prettier & ESLint in Monorepo
Previewer, ESLint, and TypeScript configurations. ESLint configurations are edited to include the simple-import-sort plugin. Prettier is configured with several options, such as tab width and trailing comma. TypeScript is temporarily removed, as no packages are using it yet. Configurations are organized into separate files for each project type: next, node, and library. Prettier and ESLint are configured to work correctly in the development environment.