Creating authentication package
Creating a package to define the application's permissions, sharing this part between the front-end and back-end, and defining the permissions in the application's code files. Configuring internal dependencies, such as eslint and prettier, and adjusting TypeScript settings. Organizing TypeScript files within the package to facilitate importing into React and Node projects.