Creating subject typing
Configuring user permissions in the project using CASL. Correctly defining the entities and actions allowed for each of them, avoiding confusion and errors in the configuration. Creating folders to organize the entities, defining the actions allowed for each one and importing these settings into the main application file. Paying attention to the order in which the entities and actions are defined to avoid bugs.